# backend/swagger.yaml
openapi: 3.0.3
info:
  title: MyActivities API
  description: |
    API REST de la plateforme **MyActivities** — découverte, planification et évaluation d'activités locales.

    ## Authentification
    L'API utilise des **JWT Bearer tokens** (access token 15 min + refresh token 7 jours).
    Les endpoints protégés nécessitent le header : `Authorization: Bearer <access_token>`

    ## Rôles utilisateurs
    | Rôle | Description |
    |------|-------------|
    | `visitor` | Non connecté — lecture seule |
    | `member` | Compte créé — planning, avis |
    | `organizer` | Référence ses activités |
    | `admin` | Modération et administration |

    ## Règles métier importantes
    - **1 avis par utilisateur par activité** (enforced côté backend)
    - **Toute nouvelle activité requiert une validation admin** avant publication
    - **Brute force** : 5 tentatives échouées → blocage 15 min
    - **Géolocalisation** : consentement explicite requis (RGPD Art. 6.1.a)
    - **Suppression de compte** : efface toutes les données personnelles (RGPD Art. 17)

    ## Codes d'erreur communs
    | Code | Signification |
    |------|---------------|
    | 400 | Données invalides |
    | 401 | Token manquant ou expiré |
    | 403 | Accès refusé (rôle insuffisant) |
    | 404 | Ressource introuvable |
    | 409 | Conflit (ex : email déjà utilisé) |
    | 429 | Trop de requêtes (rate limiting) |

  version: 1.0.0
  contact:
    name: Charles Deschamps
...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT (expiration 15 min)

  # ─────────────────────────────────────────────
  # SCHEMAS
  # ─────────────────────────────────────────────
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 3

    # ── Errors ──────────────────────────────────
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Le champ email est invalide.
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # ── Auth ────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: john.doe@example.com
        pseudo:
          type: string
          example: JohnDoe42
        role:
          type: string
          enum: [member, organizer, admin]
          example: member
        status:
          type: string
          enum: [active, suspended, deleted]
          example: active
        siret:
          type: string
          nullable: true
          description: SIRET de l'organisateur (14 chiffres) ; null pour un membre.
          example: "73282932000074"
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: https://cdn.myactivities.app/avatars/uuid.jpg
        createdAt:
          type: string
          format: date-time
        emailVerified:
          type: boolean
          example: true
    UpdateUserRequest:
      type: object
      properties:
        pseudo:
          type: string
          minLength: 3
          maxLength: 30
        avatarUrl:
          type: string
          format: uri
          nullable: true
    DeleteAccountRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          format: password
          description: Confirmation du mot de passe avant suppression définitive

    # ── Activity ─────────────────────────────────
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        author:
          type: object
          properties:
            id:
              type: string
              format: uuid
            pseudo:
              type: string
            avatarUrl:
              type: string
              format: uri
              nullable: true
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true

    # ── Messaging ───────────────────────────────
...
    ConversationParticipant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pseudo:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
        role:
          type: string
          enum: [admin, member]
...
    CreateReviewRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
    UpdateReviewRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5

    # ── Admin ─────────────────────────────────────
...
  parameters:
    activityId:
      name: activityId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    reviewId:
      name: reviewId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # ─────────────────────────────────────────────
  # RESPONSES
  # ─────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Token manquant ou expiré
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Token JWT manquant ou expiré.
    Forbidden:
      description: Rôle insuffisant pour cette action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: FORBIDDEN
            message: Accès réservé aux administrateurs.
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: Ressource introuvable.
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Obtenir son profil
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profil utilisateur connecté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags: [Users]
      summary: Mettre à jour son profil
      operationId: updateMe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Profil mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Pseudo déjà utilisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: PSEUDO_ALREADY_EXISTS
                message: Ce pseudo est déjà utilisé.

    delete:
      tags: [Users]
      summary: Supprimer son compte
      description: |
        Supprime définitivement le compte et toutes les données associées (RGPD Art. 17).
        Données effacées : compte, planning, avis, données personnelles.
        Les avis peuvent être conservés sous forme anonymisée (à documenter dans la politique de confidentialité).
      operationId: deleteMe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccountRequest'
      responses:
        '204':
          description: Compte supprimé — email de confirmation envoyé
        '401':
          description: Mot de passe de confirmation incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ══════════════════════════════
  # ACTIVITIES
  # ══════════════════════════════
...
  /activities/{activityId}/reviews:
    get:
      tags: [Reviews]
      summary: Lister les avis d'une activité
      description: Accessible sans authentification. Triés par date décroissante par défaut.
      operationId: listReviews
      parameters:
        - $ref: '#/components/parameters/activityId'
        - name: sort
          in: query
          schema:
            type: string
            enum: [date_desc, date_asc, rating_desc, rating_asc]
            default: date_desc
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Liste paginée d'avis
          content:
            application/json:
              schema:
                type: object
                properties:
                  avgRating:
                    type: number
                    format: float
                    example: 4.2
                  reviewCount:
                    type: integer
                    example: 37
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Reviews]
      summary: Déposer un avis
      description: |
        **Règles métier** (BF-029) :
        - 1 avis maximum par utilisateur par activité.
        - L'avis est réservé aux utilisateurs ayant ajouté l'activité à leur planning (anti-spam).
        Si un avis existe déjà, utiliser `PATCH /activities/{activityId}/reviews/{reviewId}`.
        La note moyenne de l'activité est recalculée immédiatement après dépôt.
      operationId: createReview
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/activityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Avis publié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: L'activité n'est pas dans le planning de l'utilisateur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: PLANNING_REQUIRED
                message: Vous devez avoir ajouté cette activité à votre planning pour la noter.
        '404':
          description: Activité introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Un avis de cet utilisateur existe déjà pour cette activité
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: REVIEW_ALREADY_EXISTS
                message: Vous avez déjà déposé un avis pour cette activité.
  /activities/{activityId}/reviews/{reviewId}:
    patch:
      tags: [Reviews]
      summary: Modifier son avis
      description: Réservé à l'auteur de l'avis. La note moyenne de l'activité est recalculée.
      operationId: updateReview
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/activityId'
        - $ref: '#/components/parameters/reviewId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      responses:
        '200':
          description: Avis mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: L'avis n'appartient pas à l'utilisateur connecté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Reviews]
      summary: Supprimer son avis
      description: Réservé à l'auteur de l'avis ou à un admin.
      operationId: deleteReview
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/activityId'
        - $ref: '#/components/parameters/reviewId'
      responses:
        '204':
          description: Avis supprimé
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ══════════════════════════════
  # ORGANIZERS
  # ══════════════════════════════
...
  /users/search:
    get:
      tags: [Friends]
      summary: Rechercher des utilisateurs
      description: Recherche d'utilisateurs actifs par pseudo (pour ajouter des amis). Exclut l'utilisateur courant.
      operationId: searchUsers
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 60 }
      responses:
        '200':
          description: Utilisateurs correspondants
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConversationParticipant'
        '400':
          description: Requête manquante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
...